[BUGFIX] Fix sgom_get_table command
According to DCR-1128 each byte in the sgom table contains 2 values
for 2 letters such that the 4 MSB are for the odd indices,
and the 4 LSB are for the even indices.
We're printing the values in the opposite way. Fix it.

type=bugfix
ticket=none
fixes=I8a90e8832b75ca62e1057b847f499c449d86b6fd

Change-Id: Ic324d62866c981f6cc4899df62f10a59edfb8380
Signed-off-by: Miri Korenblit <miriam.rachel.korenblit@intel.com>
Reviewed-on: https://git-amr-3.devtools.intel.com/gerrit/330260
Reviewed-by: Johannes Berg <johannes.berg@intel.com>
Tested-by: ec ger unix iil jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>